requirements.txt
streamlit
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
python-dotenv